added access to users tasks

+++ src/controller/task.ts
@@ -8,6 +8,7 @@ const ADD_TASK: RequestHandler = async (req, res) => {
       id: uuidv4(),
       task: req.body.task,
       status: false,
+      userId: req.body.userId,
     };
     const newTask = new TaskModel(task);
     await newTask.save();
@@ -25,6 +26,17 @@ const ADD_TASK: RequestHandler = async (req, res) => {
 };
 const DELETE_TASK: RequestHandler = async (req, res) => {
   try {
+    const taskToDel = await TaskModel.findOne({ id: req.params.id });
+    if (!taskToDel) {
+      res.status(404).json({ message: "No such task exists!" });
+      return;
+    }
+    if (taskToDel?.userId !== req.body.userId) {
+      res
+        .status(403)
+        .json({ message: "You don't have the access to this data!" });
+      return;
+    }
     const response = await TaskModel.findOneAndDelete({ id: req.params.id });
     const tasks = await TaskModel.find();
     if (response) {
@@ -33,8 +45,6 @@ const DELETE_TASK: RequestHandler = async (req, res) => {
         .json({ message: "Task successfully deleted!", tasks: tasks });
       return;
     }
-    res.status(404).json({ message: "No such task exists!" });
-    return;
   } catch (err) {
     res.status(500).json({ message: "Something went wrong!" });
     return;
@@ -52,6 +62,17 @@ const GET_TASKS: RequestHandler = async (req, res) => {
 };
 const UPDATE_TASK: RequestHandler = async (req, res) => {
   try {
+    const taskToDel = await TaskModel.findOne({ id: req.params.id });
+    if (!taskToDel) {
+      res.status(404).json({ message: "No such task exists!" });
+      return;
+    }
+    if (taskToDel?.userId !== req.body.userId) {
+      res
+        .status(403)
+        .json({ message: "You don't have the access to this data!" });
+      return;
+    }
     const taskToUpdate = await TaskModel.findOneAndUpdate(
       { id: req.params.id },
       { ...req.body },
@@ -63,7 +84,6 @@ const UPDATE_TASK: RequestHandler = async (req, res) => {
         .json({ message: "updated successfully!", task: taskToUpdate });
       return;
     }
-    res.status(404).json({ message: "No such task exists!" });
   } catch (err) {
     res.status(500).json({ message: "Something went wrong!" });
   }
